+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 13.0.1
+
+ * Begin transition back to MQTT from AMQP (ping channel only)
+
+# 13.0.0
+
+ * Nerves system upgrade
+
 # 12.3.4
 
  * Bug fix for self-hosted image uploads.